refactor: Split out Logos into component
